Responsible deletions

+++ cmd/lxdk/delete.go
@@ -1,8 +1,10 @@
 package main
 
 import (
+	"log"
 	"os"
 	"path"
+	"strings"
 
 	"github.com/greymatter-io/lxdk/config"
 	"github.com/greymatter-io/lxdk/containers"
@@ -54,7 +56,10 @@ func doDelete(ctx *cli.Context) error {
 		return err
 	}
 
-	if ctx.Bool("delete-storage") {
+	if !strings.Contains(state.StoragePool, "lxdk") {
+		log.Default().Printf("storage pool %s was not created by lxdk and will not be deleted", state.StoragePool)
+	}
+	if ctx.Bool("delete-storage") && strings.Contains(state.StoragePool, "lxdk") {
 		err = deleteStoragePool(state, is)
 		if err != nil {
 			return err
